pinakAndDjango/requirements.txt
Django>=4.2.0,<5.3.0
Pillow>=10.0.0
django-tailwind>=3.6.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
django-browser-reload>=1.11.0
python-dotenv>=1.0.0
requests>=2.31.0
django-cors-headers>=4.3.1 